adding interface for functions + cleanup
