+++ cantor-http-server/src/main/java/com/salesforce/cantor/http/jersey/EmbeddedHttpServer.java
@@ -1,7 +1,7 @@
 package com.salesforce.cantor.http.jersey;
 
 import com.salesforce.cantor.Cantor;
-import com.salesforce.cantor.functions.Functions;
+import com.salesforce.cantor.functions.FunctionsOnCantor;
 import com.salesforce.cantor.h2.CantorOnH2;
 import com.salesforce.cantor.h2.H2DataSourceProperties;
 import com.salesforce.cantor.h2.H2DataSourceProvider;
@@ -39,7 +39,7 @@ public class EmbeddedHttpServer {
                 bind(new ObjectsResource(cantor));
                 bind(new SetsResource(cantor));
                 bind(new MapsResource(cantor));
-                bind(new FunctionsResource(cantor, new Functions(cantor)));
+                bind(new FunctionsResource(cantor, new FunctionsOnCantor(cantor)));
             }
         });
